# packages/contracts/openapi/api-server.yaml
openapi: 3.1.0
info:
  title: WEOUC API Server Baseline
  version: 0.2.1
  description: 校园综合应用后端当前可联调接口契约，覆盖系统接口与微信小程序主链路基础能力。
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/student:
    get:
      tags: [IAM]
      summary: 获取当前用户资料
      operationId: getCurrentStudent
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 用户资料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentProfileResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
    post:
      tags: [IAM]
      summary: 提交教务绑定
      operationId: bindStudent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindStudentRequest'
      responses:
        '200':
          description: 绑定成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentProfileResponseEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
    put:
      tags: [IAM]
      summary: 解绑教务
      operationId: unbindStudent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStudentRequest'
      responses:
        '200':
          description: 解绑成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObjectResponseEnvelope'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
...
  responses:
    BadRequestResponse:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseEnvelope'
    UnauthorizedResponse:
      description: 未登录或登录状态无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseEnvelope'
...
    NotFoundResponse:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseEnvelope'
  schemas:
    RequestEnvelope:
      type: object
      required: [request_id]
      properties:
        request_id:
          type: string
    GenericData:
      type: object
      additionalProperties: true
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/RequestEnvelope'
        - type: object
          required: [error]
          properties:
            error:
              $ref: '#/components/schemas/ErrorDetail'
    GenericObjectResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/RequestEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/GenericData'
...
    StudentProfile:
      type: object
      required: [name, avatar_url, student_id, major, college, grade, is_bound, updated_at]
      properties:
        name:
          type: string
        avatar_url:
          type: string
        student_id:
          type: string
        major:
          type: string
        college:
          type: string
        grade:
          type: string
        is_bound:
          type: boolean
        updated_at:
          type: string
          format: date-time
    StudentProfileResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/RequestEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/StudentProfile'
...
    BindStudentRequest:
      type: object
      required: [student_id, password, captcha]
      properties:
        student_id:
          type: string
        password:
          type: string
        captcha:
          type: string
    UpdateStudentRequest:
      type: object
      properties:
        is_bound:
          type: boolean
          enum: [false]
...